move sessions api file out of src/api/ - gets regenerated from config
