+++ ui/desktop/src/components/SessionListView.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { ViewConfig } from '../App';
 import { MessageSquare, Loader, AlertCircle, Calendar, ChevronRight } from 'lucide-react';
-import { fetchSessions, type Session } from '../api/sessions';
+import { fetchSessions, type Session } from '../sessions';
 import { Card } from './ui/card';
 import { Button } from './ui/button';
 import BackButton from './ui/BackButton';
